+++ module/ptu.js
@@ -36,7 +36,7 @@ export let log = (...args) => console.log("FVTT PTU | ", ...args);
 export let warn = (...args) => console.warn("FVTT PTU | ", ...args);
 export let error = (...args) => console.error("FVTT PTU | ", ...args)
 
-export const LATEST_VERSION = "1.1.3";
+export const LATEST_VERSION = "1.1.4";
 
 Hooks.once('init', async function() {